+++ ssstar/src/lib.rs
@@ -3,10 +3,10 @@
 mod async_bridge;
 mod config;
 mod create;
+mod custom_credentials_provider;
 mod error;
 mod extract;
 mod objstore;
-mod role_credentials_provider;
 #[cfg(feature = "storage")]
 mod storage;
 mod tar;
@@ -14,6 +14,9 @@ mod writers;
 
 pub use config::Config;
 pub use create::*;
+pub use custom_credentials_provider::{
+    CustomCredentialsProvider, CustomCredentialsUpdateCallback, CustomCredentialsUpdateOutput,
+};
 pub use error::{Result, S3TarError};
 pub use extract::*;
 #[cfg(feature = "storage")]