+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-* Add `aws-sdk-ssm v0.27.0` as a dependency for `ssstar` crate
-* Add additional options to support bastion-based assume role with auto refreshing credentials
+* Add `CustomCredentialsProvider` trait to support custom way to provide credentials
+
+  The code that uses this crate as a lib can implement this trait and pass it into config
 
 ## 0.5.1 - 23-Jun-2023